requirements.txt
python-dotenv
pyjwt
cryptography
requests